+++ src/main/java/insane96mcp/iguanatweaksreborn/data/generator/client/ITRItemModelsProvider.java
@@ -3,6 +3,7 @@ package insane96mcp.iguanatweaksreborn.data.generator.client;
 import insane96mcp.iguanatweaksreborn.IguanaTweaksReborn;
 import insane96mcp.iguanatweaksreborn.module.farming.crops.Crops;
 import insane96mcp.iguanatweaksreborn.module.world.Berries;
+import insane96mcp.iguanatweaksreborn.module.world.BiomeCompass;
 import insane96mcp.iguanatweaksreborn.module.world.CyanFlower;
 import net.minecraft.data.PackOutput;
 import net.minecraft.resources.ResourceLocation;
@@ -31,6 +32,8 @@ public class ITRItemModelsProvider extends ItemModelProvider {
 
         basicItemWithTexture(CyanFlower.FLOWER.item().get(), new ResourceLocation(IguanaTweaksReborn.MOD_ID, "block/cyan_flower"));
         basicItemWithTexture(Crops.SOLANUM_NEOROSSII.item().get(), new ResourceLocation(IguanaTweaksReborn.MOD_ID, "block/solanum_neorossii"));
+
+        basicItem(BiomeCompass.COMPASS.get());
     }
     public ItemModelBuilder basicItemWithTexture(Item item, ResourceLocation texture)
     {